backend/requirements.txt
mysql-connector-python
pyjwt
bcrypt
reportlab
python-dotenv
